+++ performance_test/src/experiment_execution/data_entity_runner.cpp
@@ -16,10 +16,10 @@
 
 #include <memory>
 
-#include "performance_test/communication_abstractions/communicator_factory.hpp"
 #include "performance_test/execution_tasks/publisher_task.hpp"
 #include "performance_test/execution_tasks/subscriber_task.hpp"
 #include "performance_test/experiment_configuration/experiment_configuration.hpp"
+#include "performance_test/experiment_execution/pub_sub_factory.hpp"
 #include "performance_test/experiment_execution/runner.hpp"
 
 namespace performance_test
@@ -32,14 +32,14 @@ DataEntityRunner::DataEntityRunner(const ExperimentConfiguration & ec)
       std::make_shared<SubscriberTask>(
         ec,
         m_sub_stats.at(i),
-        CommunicatorFactory::get_subscriber(ec)));
+        PubSubFactory::get().create_subscriber(ec)));
   }
   for (uint32_t i = 0; i < m_ec.number_of_publishers; ++i) {
     m_pubs.push_back(
       std::make_shared<PublisherTask>(
         ec,
         m_pub_stats.at(i),
-        CommunicatorFactory::get_publisher(ec)));
+        PubSubFactory::get().create_publisher(ec)));
   }
 }